Move mode logic into mode class. Add mode methods usage in the game class
